WeirdMatrices

WeirdMatrices.intersect_with_indices

WeirdMatrices.intersect_with_indices — Method

intersect_with_indices(u, v)

Intersect u and v and return the elements present in both, as well as their indices in u and v, respectively.

Credits: @MasonProtter